+++ holberton.h
@@ -3,5 +3,7 @@
 
 int _putchar(char c);
 int _printf(const char *format, ...);
+void printnumber(int n);
+void printUn(unsigned int n);
 #endif
